update wishlist functionality
